Clean up install.sh

+++ install.sh
@@ -16,7 +16,7 @@ message() {
 
     echo -e "\e$color"
     msg="# $2 #"
-    edge=$(echo $msg | sed 's/./#/g')
+    edge=$(echo "$msg" | sed 's/./#/g')
     echo -e "$edge\n$msg\n$edge"
     echo -e "\e[0m"
 }
@@ -24,24 +24,24 @@ message() {
 message yellow "Updating package repositories..."
 sudo apt update -y
 
-cat $BASEDIR/packages.list | while read package; do
-    if [ -z "$(dpkg -s $package)" ] ; then
+while read -r package; do
+    if [ -z "$(dpkg -s "$package")" ] ; then
         message yellow "Installing $package..." 
-        sudo apt install -y $package
+        sudo apt install -y "$package"
     else
         message green "$package already installed!"
     fi
-done
+done < "$BASEDIR/packages.list"
 
 message yellow "Installing oh-my-zsh..."a
 if [ ! -d "$HOME/.oh-my-zsh/" ] ; then
-    git clone https://github.com/robbyrussell/oh-my-zsh.git $HOME/.oh-my-zsh
+    git clone https://github.com/robbyrussell/oh-my-zsh.git "$HOME/.oh-my-zsh"
 fi
 
 echo "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions"
 
 if [ ! -d "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/plugins/zsh-autosuggestions" ] ; then
-    git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+    git clone https://github.com/zsh-users/zsh-autosuggestions "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions"
 fi
 
 message yellow "Installing pywal..."
@@ -53,9 +53,9 @@ message yellow "Installing Vundle for Vim..."
 if [ ! -d "$HOME/.vim/bundle/Vundle.vim" ] ; then
     git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 else 
-    cd $HOME/.vim/bundle/Vundle.vim
+    cd "$HOME/.vim/bundle/Vundle.vim"
     git pull
-    cd $BASEDIR
+    cd "$BASEDIR"
 fi
 
 vim +PluginInstall +qall
@@ -63,33 +63,33 @@ vim +PluginInstall +qall
 if [ -z "$(command -v polybar)" ] ; then
     message yellow "Installing Polybar..."
     if [ ! -d "$BASEDIR/polybar" ] ; then
-        git clone --branch 3.2 --recursive https://github.com/jaagr/polybar $BASEDIR/polybar
+        git clone --branch 3.2 --recursive https://github.com/jaagr/polybar "$BASEDIR/polybar"
     else
-        cd $BASEDIR/polybar
+        cd "$BASEDIR/polybar"
         git pull
-        cd $BASEDIR
+        cd "$BASEDIR"
     fi
-    mkdir -p $BASEDIR/polybar/build
-    cd $BASEDIR/polybar/build
-    cmake $BASEDIR/polybar
+    mkdir -p "$BASEDIR/polybar/build"
+    cd "$BASEDIR/polybar/build"
+    cmake "$BASEDIR/polybar"
     sudo make install
-    cd $BASEDIR
+    cd "$BASEDIR"
 else
     message green "Polybar already installed!"
 fi
 
-rm -r $BASEDIR/polybar
+rm -r "$BASEDIR/polybar"
 
 if [ -z "$(command -v i3)" ] ; then
     message yellow "Installing i3-gaps..."
     if [ ! -d "$BASEDIR/i3-gaps" ] ; then
         git clone https://www.github.com/Airblader/i3 i3-gaps
     else
-        cd $BASEDIR/i3-gaps
+        cd "$BASEDIR/i3-gaps"
         git pull
-        cd $BASEDIR
+        cd "$BASEDIR"
     fi
-    cd $BASEDIR/i3-gaps
+    cd "$BASEDIR/i3-gaps"
     autoreconf --force --install
     rm -r build/
     mkdir -p build && cd build/
@@ -100,11 +100,11 @@ else
     message green "i3-gaps already installed!"
 fi
 
-rm -r $BASEDIR/i3-gaps
+rm -r "$BASEDIR/i3-gaps"
 
-cd $BASEDIR
+cd "$BASEDIR"
 
-if [ -z "$(fc-list | grep "Hack")" ] ; then
+if fc-list | ! grep -q 'Hack'  ; then
     message yellow "Installing fonts..."
     wget -O fonts.zip https://github.com/ryanoasis/nerd-fonts/releases/download/v2.0.0/Hack.zip
     mkdir fonts
@@ -112,7 +112,7 @@ if [ -z "$(fc-list | grep "Hack")" ] ; then
     cp -R fonts/ ~/.local/share/fonts/
     rm -r fonts/
     fc-cache -f -v
-    fc-list } grep "Hack"
+    fc-list | grep "Hack"
 else
     message green "fonts already instaleed!"
 fi
@@ -134,20 +134,20 @@ Files=(
 
 for i in "${Files[@]:0:3}"; do
     echo "rm $HOME/$i"
-    rm $HOME/$i
+    rm "${HOME:?}/$i"
 	echo "ln -srf $BASEDIR/$i $HOME/$i"
-	ln -srf $BASEDIR/$i $HOME/$i
+	ln -srf "$BASEDIR/$i" "$HOME/$i"
 done
 
 for i in "${Files[@]:3:7}"; do
     echo "rm -r $HOME/$i"
-    rm -r $HOME/$i
+    rm -r "${HOME:?}/$i"
     echo "ln -srf $BASEDIR/$i $HOME/.config/"
-    ln -srf $BASEDIR/$i $HOME/.config/
+    ln -srf "$BASEDIR/$i" "$HOME/.config/"
 done
 
 message yellow "Applying theme..."
-wal -n -a "93" -i "$BASEDIR/outrun.png"
+wal -n -i "$BASEDIR/mountain.jpg"
 
 message yellow "Changing shell..."
 chsh -s /bin/zsh
